Moved pool image upload to new component
