+++ services/client/src/pages/AddHypePool/DetailsForm/DetailsForm.tsx
@@ -14,34 +14,23 @@ import {
   InfoCard,
   FormAction,
   FormElement,
-  PoolImage,
 } from '../AddHypePool.styled';
 import TextArea from '../../../components/textarea/TextArea';
 import Box from '../../../components/styles/Box';
-import { UploadControl } from '../../../components/upload/Upload';
-import LoadingSpinner from '../../../assets/icons/Spinner';
-import SuccessIcon from '../../../assets/icons/Success';
+import { HypeImageUpload } from './HypeImageUpload';
 
 export interface DetailsFormProps {
   defaultValues: HypePoolDetailsForm;
   onSubmit: (data: HypePoolDetailsForm) => void;
-  onUploadImage: () => void;
-  removeImage: () => void;
-  selectedImage: File;
-  setSelectedImage: Dispatch<any>;
-  isUploadingImage: boolean;
+  setImageUrl: Dispatch<any>;
   imageUrl: string;
 }
 
 export const DetailsForm = ({
   defaultValues,
   onSubmit,
-  onUploadImage,
-  removeImage,
-  selectedImage,
-  setSelectedImage,
-  isUploadingImage,
   imageUrl,
+  setImageUrl,
 }: DetailsFormProps) => {
   const { register, handleSubmit, errors, authenticated } = useDetailsFormEffects(defaultValues);
 
@@ -107,65 +96,7 @@ export const DetailsForm = ({
           />
         </FormElement>
 
-        <FormElement>
-          <Box display="flex" flexDirection="row" gridGap="0.2rem" alignItems="center">
-            <Label>Project's image: </Label>
-          </Box>
-          <Example>
-            This is not mandatory but we recommend uploading an image that represents your project
-          </Example>
-
-          {selectedImage && (
-            <Box display="flex" flexDirection="column" gridGap="0.2rem" mb={3}>
-              <PoolImage alt="not found" width={'250px'} src={URL.createObjectURL(selectedImage)} />
-              <Button
-                size="full-width"
-                disabled={isUploadingImage}
-                type="button"
-                variant="primary"
-                onClick={removeImage}
-              >
-                Remove
-              </Button>
-            </Box>
-          )}
-          {!selectedImage && (
-            <Button type="button" style={{ padding: 0 }} variant="primary">
-              <UploadControl
-                value={selectedImage}
-                onChange={(event) => {
-                  setSelectedImage(event.target.files[0]);
-                }}
-                accept="image/*"
-              >
-                Select image
-              </UploadControl>
-            </Button>
-          )}
-          {selectedImage &&
-            (!isUploadingImage ? (
-              <>
-                <Button
-                  type="button"
-                  variant="primary"
-                  onClick={() => onUploadImage()}
-                  disabled={!!imageUrl}
-                >
-                  {!imageUrl ? (
-                    'Upload image'
-                  ) : (
-                    <Box display="flex" alignItems="center">
-                      <Text pr={3}>Image uploaded</Text> <SuccessIcon />
-                    </Box>
-                  )}
-                </Button>
-              </>
-            ) : (
-              <Box p={3}>
-                <LoadingSpinner />
-              </Box>
-            ))}
-        </FormElement>
+        <HypeImageUpload imageUrl={imageUrl} setImageUrl={setImageUrl} />
 
         {/* Description */}
         <FormElement>
